@@ -1,6 +1,6 @@
 {
   "name": "custom-functions-metadata-plugin",
-  "version": "0.0.3",
+  "version": "1.2.4",
   "author": "Office Dev",
   "bugs": {
     "url": "https://github.com/OfficeDev/Office-Addin-Scripts/issues"
@@ -18,22 +18,25 @@
     "url": "https://github.com/OfficeDev/Office-Addin-Scripts"
   },
   "scripts": {
-    "tsc": "tsc -p tsconfig.json",
-    "tsc-watch": "tsc -p tsconfig.json -w",
-    "build": "rimraf lib && concurrently \"npm run tsc\"",
-    "test": "echo \"No tests.\""
+    "build": "rimraf lib && concurrently \"tsc -p tsconfig.json\"",
+    "lint": "office-addin-lint check",
+    "lint:fix": "office-addin-lint fix",
+    "prettier": "office-addin-lint prettier",
+    "test": "echo \"No tests.\"",
+    "watch": "tsc -p tsconfig.json -w"
   },
   "dependencies": {
-    "custom-functions-metadata": "^0.0.3",
-    "es6-promise": "^4.2.5",
-    "rimraf": "^2.6.2",
-    "typescript": "^3.1.6",
-    "webpack": "^4.27.1"
+    "custom-functions-metadata": "^1.2.4",
+    "loader-utils": "^2.0.0",
+    "webpack": "^5.24.4"
   },
   "devDependencies": {
-    "@types/node": "^10.12.12",
-    "@types/webpack": "^4.4.21",
-    "concurrently": "^4.1.0",
-    "tslint": "^5.11.0"
-  }
+    "@types/loader-utils": "^2.0.2",
+    "@types/node": "^14.17.2",
+    "concurrently": "^6.2.2",
+    "office-addin-lint": "^1.4.4",
+    "rimraf": "^3.0.2",
+    "typescript": "^4.3.5"
+  },
+  "gitHead": "dffbcd4d1397534105b98fb13ba0fb1c3aba2dea"
 }
