@@ -1,20 +1,23 @@
 {
   "name": "custom-functions-metadata",
-  "version": "0.0.3",
+  "version": "1.2.4",
   "description": "Generate metadata for Excel Custom Functions.",
-  "main": "./lib/custom-functions-metadata.js",
+  "main": "./lib/main.js",
   "scripts": {
-    "tsc": "tsc -p tsconfig.json",
-    "tsc-watch": "tsc -p tsconfig.json -w",
-    "build": "rimraf lib && concurrently \"npm run tsc\"",
-    "start": "rimraf lib && concurrently \"npm run tsc-watch\"",
-    "test": "mocha -r ts-node/register test/src/test.ts",
-    "generate-json": "node ./lib/json-generate.js generate"
+    "build": "rimraf lib && concurrently \"tsc -p tsconfig.json\"",
+    "cli": "node lib/cli.js",
+    "lint": "office-addin-lint check",
+    "lint:fix": "office-addin-lint fix",
+    "package": "node ./lib/cli.js generate",
+    "prettier": "office-addin-lint prettier",
+    "start": "rimraf lib && concurrently \"npm run watch\"",
+    "test": "mocha -r ts-node/register test/src/*.ts",
+    "watch": "tsc -p tsconfig.json -w"
   },
   "author": "Office Dev",
   "license": "MIT",
   "bin": {
-    "custom-functions-metadata": "./lib/json-generate.js"
+    "custom-functions-metadata": "./cli.js"
   },
   "keywords": [
     "Office",
@@ -22,27 +25,25 @@
     "Excel Custom Funtions"
   ],
   "dependencies": {
-    "@types/assert": "^1.4.0",
-    "@types/xregexp": "^3.0.29",
-    "assert": "^1.4.1",
-    "commander": "^2.19.0",
-    "fs": "0.0.1-security",
-    "nconf": "^0.10.0",
-    "optimist": "^0.6.1",
-    "xregexp": "^4.2.4"
+    "assert": "^1.5.0",
+    "commander": "^6.2.0",
+    "nconf": "^0.11.3",
+    "office-addin-cli": "^1.3.3",
+    "xregexp": "^4.3.0"
   },
   "devDependencies": {
-    "@types/commander": "^2.12.2",
-    "@types/es6-promise": "^3.3.0",
-    "@types/mocha": "^5.2.5",
-    "@types/node": "^10.12.0",
-    "@types/node-fetch": "^2.1.2",
-    "concurrently": "^3.6.1",
-    "mocha": "^5.2.0",
-    "rimraf": "^2.6.2",
-    "ts-node": "^7.0.1",
-    "tslint": "^5.11.0",
-    "typescript": "^3.1.3"
+    "@types/assert": "^1.4.6",
+    "@types/mocha": "^7.0.2",
+    "@types/node": "^14.17.2",
+    "@types/node-fetch": "^2.5.10",
+    "@types/xregexp": "^3.0.30",
+    "concurrently": "^6.2.2",
+    "mocha": "^9.1.1",
+    "office-addin-lint": "^1.4.4",
+    "office-addin-prettier-config": "^1.1.2",
+    "rimraf": "^3.0.2",
+    "ts-node": "^10.1.0",
+    "typescript": "^4.3.5"
   },
   "repository": {
     "type": "git",
@@ -50,5 +51,7 @@
   },
   "bugs": {
     "url": "https://github.com/OfficeDev/Office-Addin-Scripts/issues"
-  }
+  },
+  "prettier": "office-addin-prettier-config",
+  "gitHead": "dffbcd4d1397534105b98fb13ba0fb1c3aba2dea"
 }
