@@ -1,47 +1,52 @@
 {
   "name": "office-addin-debugging",
-  "version": "2.0.0",
+  "version": "4.3.5",
   "description": "For debugging Office Add-ins.",
-  "main": "./lib/debugging.js",
+  "main": "./lib/main.js",
   "scripts": {
-    "tsc": "tsc -p tsconfig.json",
-    "tsc-watch": "tsc -p tsconfig.json -w",
-    "build": "rimraf lib && concurrently \"npm run tsc\"",
-    "start": "rimraf lib && concurrently \"npm run tsc-watch\"",
+    "build": "rimraf lib && concurrently \"tsc -p tsconfig.json\"",
+    "cli": "node lib/cli.js",
+    "lint": "office-addin-lint check",
+    "lint:fix": "office-addin-lint fix",
+    "node-debugger": "office-addin-node-debugger",
+    "prettier": "office-addin-lint prettier",
+    "start": "rimraf lib && concurrently \"npm run watch\"",
     "test": "mocha -r ts-node/register test/**/*.ts",
-    "node-debugger": "office-addin-node-debugger"
+    "watch": "tsc -p tsconfig.json -w"
   },
   "author": "Office Dev",
   "license": "MIT",
   "bin": {
-    "office-addin-debugging": "./lib/debugging.js"
+    "office-addin-debugging": "./cli.js"
   },
   "keywords": [
     "Office"
   ],
   "dependencies": {
     "child_process": "^1.0.2",
-    "commander": "^2.19.0",
-    "node-fetch": "^2.2.0",
-    "office-addin-dev-settings": "^1.4.0",
-    "office-addin-manifest": "^1.2.1",
-    "office-addin-node-debugger": "^0.3.2"
+    "commander": "^6.2.0",
+    "node-fetch": "^2.6.1",
+    "office-addin-cli": "^1.3.3",
+    "office-addin-dev-certs": "^1.7.4",
+    "office-addin-dev-settings": "^1.10.5",
+    "office-addin-manifest": "^1.7.5",
+    "office-addin-node-debugger": "^0.6.4",
+    "office-addin-usage-data": "^1.4.4"
   },
   "devDependencies": {
-    "@types/es6-promise": "0.0.32",
-    "@types/express": "^4.16.0",
+    "@types/express": "^4.17.3",
     "@types/mocha": "^5.2.5",
-    "@types/node": "^10.12.0",
-    "@types/node-fetch": "^2.1.2",
-    "@types/ws": "^6.0.1",
-    "concurrently": "^3.5.0",
+    "@types/node": "^14.17.2",
+    "@types/node-fetch": "^2.5.10",
+    "@types/ws": "^6.0.4",
+    "concurrently": "^6.2.2",
     "express": "^4.16.4",
-    "mocha": "^5.2.0",
-    "rimraf": "^2.6.1",
-    "ts-node": "^7.0.1",
-    "tslint": "^5.11.0",
-    "typescript": "^3.1.3",
-    "ws": "^6.1.0"
+    "mocha": "^9.1.1",
+    "office-addin-lint": "^1.4.4",
+    "rimraf": "^2.7.1",
+    "ts-node": "^10.1.0",
+    "typescript": "^4.3.5",
+    "ws": "^7.4.6"
   },
   "repository": {
     "type": "git",
@@ -49,5 +54,7 @@
   },
   "bugs": {
     "url": "https://github.com/OfficeDev/Office-Addin-Scripts/issues"
-  }
+  },
+  "prettier": "office-addin-prettier-config",
+  "gitHead": "dffbcd4d1397534105b98fb13ba0fb1c3aba2dea"
 }
