@@ -1,17 +1,21 @@
 {
   "name": "office-addin-dev-settings",
-  "version": "1.4.0",
+  "version": "1.10.5",
   "description": "Configure developer settings for Office Add-ins.",
-  "main": "./lib/dev-settings.js",
+  "main": "./lib/main.js",
   "scripts": {
     "build": "rimraf lib && concurrently \"tsc -p tsconfig.json\"",
-    "watch": "rimraf lib && concurrently \"tsc -p tsconfig.json -w\"",
-    "test": "mocha -r ts-node/register test/**/*.ts"
+    "cli": "node lib/cli.js",
+    "lint": "office-addin-lint check",
+    "lint:fix": "office-addin-lint fix",
+    "prettier": "office-addin-lint prettier",
+    "test": "mocha -r ts-node/register test/**/*.ts",
+    "watch": "rimraf lib && concurrently \"tsc -p tsconfig.json -w\""
   },
   "author": "Office Dev",
   "license": "MIT",
   "bin": {
-    "office-addin-dev-settings": "./lib/dev-settings.js"
+    "office-addin-dev-settings": "./cli.js"
   },
   "keywords": [
     "Office",
@@ -19,27 +23,35 @@
   ],
   "dependencies": {
     "child_process": "^1.0.2",
-    "commander": "^2.18.0",
-    "fs": "0.0.1-security",
-    "office-addin-manifest": "^1.2.1",
-    "whatwg-url": "^7.0.0",
+    "commander": "^6.2.0",
+    "fs-extra": "^9.0.1",
+    "inquirer": "^7.3.3",
+    "jszip": "^3.7.1",
+    "junk": "^3.1.0",
+    "office-addin-cli": "^1.3.3",
+    "office-addin-manifest": "^1.7.5",
+    "office-addin-usage-data": "^1.4.4",
+    "open": "^6.4.0",
+    "whatwg-url": "^7.1.0",
     "winreg": "^1.2.4"
   },
   "devDependencies": {
-    "@types/es6-promise": "^3.3.0",
-    "@types/fs-extra": "^5.0.4",
-    "@types/mocha": "^5.2.5",
-    "@types/node": "^10.5.7",
-    "@types/node-fetch": "^2.1.2",
+    "@types/fs-extra": "^9.0.1",
+    "@types/inquirer": "^6.5.0",
+    "@types/mocha": "^8.0.3",
+    "@types/node": "^14.17.2",
+    "@types/node-fetch": "^2.5.10",
+    "@types/semver": "^7.3.4",
+    "@types/sinon": "^7.5.2",
     "@types/whatwg-url": "^6.4.0",
     "@types/winreg": "^1.2.30",
-    "concurrently": "^3.6.1",
-    "fs-extra": "^7.0.1",
-    "mocha": "^5.2.0",
-    "rimraf": "^2.6.2",
-    "ts-node": "^7.0.1",
-    "tslint": "^5.11.0",
-    "typescript": "^3.0.3"
+    "concurrently": "^6.2.2",
+    "mocha": "^9.1.1",
+    "office-addin-lint": "^1.4.4",
+    "rimraf": "^3.0.2",
+    "sinon": "^7.5.0",
+    "ts-node": "^10.1.0",
+    "typescript": "^4.3.5"
   },
   "repository": {
     "type": "git",
@@ -47,5 +59,7 @@
   },
   "bugs": {
     "url": "https://github.com/OfficeDev/Office-Addin-Scripts/issues"
-  }
+  },
+  "prettier": "office-addin-prettier-config",
+  "gitHead": "dffbcd4d1397534105b98fb13ba0fb1c3aba2dea"
 }
