@@ -1,47 +1,52 @@
 {
   "name": "office-addin-manifest",
-  "version": "1.2.1",
+  "version": "1.7.5",
   "description": "Read and modify Office Add-in manifest files.",
-  "main": "./lib/manifest.js",
+  "main": "./lib/main.js",
   "scripts": {
     "build": "rimraf lib && concurrently \"tsc -p tsconfig.json\"",
-    "watch": "rimraf lib && concurrently \"tsc -p tsconfig.json -w\"",
-    "test": "mocha -r ts-node/register test/**/*.ts"
+    "cli": "node lib/cli.js",
+    "lint": "office-addin-lint check",
+    "lint:fix": "office-addin-lint fix",
+    "prettier": "office-addin-lint prettier",
+    "test": "mocha -r ts-node/register test/**/*.ts",
+    "watch": "rimraf lib && concurrently \"tsc -p tsconfig.json -w\""
   },
   "author": "Office Dev",
   "license": "MIT",
   "bin": {
-    "office-addin-manifest": "./lib/manifest.js"
+    "office-addin-manifest": "./cli.js"
   },
   "keywords": [
     "Office",
     "Office Add-in"
   ],
   "dependencies": {
-    "commander": "^2.19.0",
-    "fs": "0.0.1-security",
-    "node-fetch": "^2.2.0",
+    "chalk": "^2.4.2",
+    "commander": "^6.2.0",
+    "node-fetch": "^2.6.1",
+    "office-addin-cli": "^1.3.3",
+    "office-addin-usage-data": "^1.4.4",
     "path": "^0.12.7",
-    "uuid": "^3.3.2",
-    "xml2js": "^0.4.19"
+    "uuid": "^8.3.2",
+    "xml2js": "^0.4.23"
   },
   "devDependencies": {
-    "@types/es6-promise": "0.0.32",
-    "@types/mocha": "^5.2.5",
-    "@types/node": "^10.12.0",
-    "@types/node-fetch": "^2.1.2",
-    "@types/uuid": "^3.4.4",
-    "@types/validator": "^9.4.2",
-    "@types/xml2js": "^0.4.3",
-    "concurrently": "^3.6.1",
+    "@types/mocha": "^5.2.7",
+    "@types/node": "^14.17.2",
+    "@types/node-fetch": "^2.5.10",
+    "@types/uuid": "^8.3.0",
+    "@types/validator": "^9.4.4",
+    "@types/xml2js": "^0.4.5",
+    "concurrently": "^6.2.2",
     "copy-dir": "^0.4.0",
     "fs-extra": "^7.0.0",
-    "mocha": "^5.2.0",
-    "rimraf": "^2.6.2",
-    "ts-node": "^7.0.1",
-    "tslint": "^5.11.0",
-    "typescript": "^3.1.3",
-    "validator": "^10.8.0"
+    "mocha": "^9.1.1",
+    "office-addin-lint": "^1.4.4",
+    "rimraf": "^3.0.2",
+    "ts-node": "^10.1.0",
+    "typescript": "^4.3.5",
+    "validator": "^13.7.0"
   },
   "repository": {
     "type": "git",
@@ -49,5 +54,7 @@
   },
   "bugs": {
     "url": "https://github.com/OfficeDev/Office-Addin-Scripts/issues"
-  }
+  },
+  "prettier": "office-addin-prettier-config",
+  "gitHead": "dffbcd4d1397534105b98fb13ba0fb1c3aba2dea"
 }
