@@ -1,23 +1,26 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
 /**
  * Test comments
  * @param {number} first - the first number
  * @param {number} second 
- * @helpUrl https://dev.office.com
- * @CustomFunction
+ * @helpUrl https://docs.microsoft.com/office/dev/add-ins
+ * @customfunction
  * @notfound test123
  * @volatile
  * @streaming
  * @cancelable
- * @return {returntypetest}
+ * @return {number}
  */
 function add(first: number, second: number): number {
     return first + second;
 }
 
 /**
- * @param {string} one - onetest
- * @param {number} x - x arraynumber
- * @CustomFunction
+ * @param {string[][]} one - onetest
+ * @param {Array<Array<number>>} x - x arraynumber
+ * @customfunction
  */
 function complexFunction(one: string[][], x: Array<Array<number>>): string[][] {
     return [""][""];
@@ -28,7 +31,7 @@ function notAdded() {
 
 /**
  * Testing boolean
- * @CustomFunction
+ * @customfunction
  */
 function testBool(abc: boolean): boolean {
     return true;
@@ -37,7 +40,7 @@ function testBool(abc: boolean): boolean {
 /**
  * Test function for number type
  * @param one - A number
- * @CustomFunction
+ * @customfunction
  */
 function testNumber(one: number): number {
     return 0;
@@ -46,7 +49,7 @@ function testNumber(one: number): number {
 /**
  * Test function for string type
  * @param word - Some string
- * @CustomFunction
+ * @customfunction
  */
 function testString(word: string): string {
     return "";
@@ -54,7 +57,7 @@ function testString(word: string): string {
 
 /**
  * Test function for void type
- * @CustomFunction
+ * @customfunction
  */
 function voidTest(): void {
 }
@@ -62,7 +65,7 @@ function voidTest(): void {
 /**
  * Test function for object type
  * @param obj - Some object
- * @CustomFunction
+ * @customfunction
  */
 function objectTest(obj: object): object {
     let o;
@@ -70,7 +73,7 @@ function objectTest(obj: object): object {
 }
 
 /**
- * @CustomFunction
+ * @customfunction
  */
 function testdatetime(d: number): string {
     return "";
@@ -81,7 +84,7 @@ enum Color {Red,Green,Blue};
 /**
  * Test function for enum type
  * @param e - enum type
- * @CustomFunction
+ * @customfunction
  */
 function enumTest(e: Color) : Color {
     let r : Color;
@@ -91,7 +94,7 @@ function enumTest(e: Color) : Color {
 /**
  * Test function for tuple type
  * @param t 
- * @CustomFunction
+ * @customfunction
  */
 function tupleTest(t:[string,number]){
 }
@@ -100,7 +103,7 @@ function tupleTest(t:[string,number]){
  * Test function for streaming type
  * @param x - Test string
  * @param sf - Streaming function type return type should be number
- * @CustomFunction 
+ * @customfunction 
  */
 function streamingTest(x: string, sf: CustomFunctions.StreamingHandler<number>) {
 }
@@ -108,14 +111,14 @@ function streamingTest(x: string, sf: CustomFunctions.StreamingHandler<number>) 
 /**
  * Test function for optional parameter
  * @param x - Optional string
- * @CustomFunction
+ * @customfunction
  */
 function testOptional(x?: string){
 }
 
 /**
  * Test any type
- * @CustomFunction
+ * @customfunction
  */
 function testAny(a: any): any {}
 
@@ -123,7 +126,7 @@ function testAny(a: any): any {}
  * Test support for the CustomFunctions.CancelableHandler
  * @param x string parameter
  * @param cf Cancelable Handler parameter
- * @CustomFunction
+ * @customfunction
  */
 async function testCancelableFunction(x: string, cf: CustomFunctions.CancelableHandler ): Promise<number> {
     return 1;
@@ -132,7 +135,7 @@ async function testCancelableFunction(x: string, cf: CustomFunctions.CancelableH
 /**
  * Test the custom function id and name
  * @param x test string
- * @CustomFunction updateId updateName
+ * @customfunction updateId updateName
  */
 function customFunctionIdNameTest(x:string){
 }
@@ -141,10 +144,40 @@ function customFunctionIdNameTest(x:string){
  * Test the requiresAddress tag
  * @param x string
  * @param handler my handler
- * @CustomFunction
+ * @customfunction
  * @requiresAddress
  */
-function requiresAddressTest(x: string, handler: CustomFunctions.StreamingHandler<number>){}
+function requiresAddressTest(x: string, handler: CustomFunctions.Invocation){}
+
+/**
+ * Test the CustomFunctions.Invocation type
+ * @customfunction
+ * @param invocation Invocation parameter
+ * @requiresAddress
+ */
+function customFunctionInvocationTest(x: string, invocation: CustomFunctions.Invocation){}
+
+/**
+ * Test the new cancelable type
+ * @param x string
+ * @param cancel CustomFunctions.CancelableInvocation type
+ * @customfunction
+ * @requiresAddress
+ */
+function customFunctionCancelableInvocationTest(x: string, cancel: CustomFunctions.CancelableInvocation){}
+
+/**
+ * Test the new streaming invocation type
+ * @param x string
+ * @param stream StreamingInvocation type
+ * @customfunction
+ */
+function customFunctionStreamingInvocationTest(x: string, stream: CustomFunctions.StreamingInvocation<string>){}
+
+/**
+ * @customfunction
+ */
+function UPPERCASE(){}
 
 CustomFunctionMappings.ADD=add;
 
