@@ -1,6 +1,12 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
 import * as childProcess from "child_process";
 import * as crypto from "crypto";
 import * as net from "net";
+import { ExpectedError } from "office-addin-usage-data";
+
+/* global process */
 
 /**
  * Determines whether a port is in use.
@@ -32,10 +38,10 @@ export function isPortInUse(port: number): Promise<boolean> {
  * @example "Local Address" returns undefined
  */
 function parsePort(text: string): number | undefined {
-    const result = text.match(/:(\d+)$/);
+  const result = text.match(/:(\d+)$/);
 
-    return result ? parseInt(result[1], 10) : undefined;
-  }
+  return result ? parseInt(result[1], 10) : undefined;
+}
 
 /**
  * Return the process ids using the port.
@@ -47,7 +53,8 @@ export function getProcessIdsForPort(port: number): Promise<number[]> {
 
   return new Promise((resolve, reject) => {
     const isWin32 = process.platform === "win32";
-    const command = isWin32 ? `netstat -ano` : `lsof -n -i:${port}`;
+    const isLinux = process.platform === "linux";
+    const command = isWin32 ? `netstat -ano` : isLinux ? `netstat -tlpna | grep :${port}` : `lsof -n -i:${port}`;
 
     childProcess.exec(command, (error, stdout) => {
       if (error) {
@@ -58,10 +65,11 @@ export function getProcessIdsForPort(port: number): Promise<number[]> {
           reject(error);
         }
       } else {
-        const processIds = new Set();
+        const processIds = new Set<number>();
         const lines = stdout.trim().split("\n");
         if (isWin32) {
           lines.forEach((line) => {
+            /* eslint-disable no-unused-vars */
             const [protocol, localAddress, foreignAddress, status, processId] = line.split(" ").filter((text) => text);
             if (processId !== undefined) {
               const localAddressPort = parsePort(localAddress);
@@ -70,10 +78,38 @@ export function getProcessIdsForPort(port: number): Promise<number[]> {
               }
             }
           });
+        } else if (isLinux) {
+          lines.forEach((line) => {
+            const [
+              proto /* eslint-disable-line no-unused-vars */,
+              recv /* eslint-disable-line no-unused-vars */,
+              send /* eslint-disable-line no-unused-vars */,
+              local_address,
+              remote_address /* eslint-disable-line no-unused-vars */,
+              state /* eslint-disable-line no-unused-vars */,
+              program,
+            ] = line.split(" ").filter((text) => text);
+            if (local_address !== undefined && local_address.endsWith(`:${port}`) && program !== undefined) {
+              const pid = parseInt(program, 10);
+              if (!isNaN(pid)) {
+                processIds.add(pid);
+              }
+            }
+          });
         } else {
           lines.forEach((line) => {
-            const [process, processId, user, fd, type, device, size, node, name] = line.split(" ").filter((text) => text);
-            if ((processId !== undefined) && (processId !== "PID")) {
+            const [
+              process /* eslint-disable-line no-unused-vars */,
+              processId,
+              user /* eslint-disable-line no-unused-vars */,
+              fd /* eslint-disable-line no-unused-vars */,
+              type /* eslint-disable-line no-unused-vars */,
+              device /* eslint-disable-line no-unused-vars */,
+              size /* eslint-disable-line no-unused-vars */,
+              node /* eslint-disable-line no-unused-vars */,
+              name /* eslint-disable-line no-unused-vars */,
+            ] = line.split(" ").filter((text) => text);
+            if (processId !== undefined && processId !== "PID") {
               processIds.add(parseInt(processId, 10));
             }
           });
@@ -102,7 +138,7 @@ export async function randomPortNotInUse(): Promise<number> {
  * Returns a random number between 0 and 65535
  */
 function randomPortNumber(): number {
-   return crypto.randomBytes(2).readUInt16LE(0, false);
+  return crypto.randomBytes(2).readUInt16LE(0);
 }
 
 /**
@@ -111,7 +147,7 @@ function randomPortNumber(): number {
  * @throws Error if port is not a number from 0 to 65535.
  */
 function validatePort(port: number): void {
-  if ((typeof(port) !== "number") || (port < 0) || (port > 65535)) {
-    throw new Error("Port should be a number from 0 to 65535.");
+  if (typeof port !== "number" || port < 0 || port > 65535) {
+    throw new ExpectedError("Port should be a number from 0 to 65535.");
   }
 }
