@@ -1,5 +1,18 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
 import * as childProcess from "child_process";
+import inquirer = require("inquirer");
+import { OfficeAddinManifest } from "office-addin-manifest";
 import { URL } from "whatwg-url";
+import { ExpectedError } from "office-addin-usage-data";
+
+/* global process */
+
+export const EdgeBrowserAppcontainerName: string = "Microsoft.MicrosoftEdge_8wekyb3d8bbwe";
+export const EdgeWebViewAppcontainerName: string = "Microsoft.win32webviewhost_cw5n1h2txyewy";
+export const EdgeBrowserName: string = "Microsoft Edge Web Browser";
+export const EdgeWebViewName: string = "Microsoft Edge WebView";
 
 /**
  * Adds a loopback exemption for the appcontainer.
@@ -8,7 +21,7 @@ import { URL } from "whatwg-url";
  */
 export function addLoopbackExemptionForAppcontainer(name: string): Promise<void> {
   if (!isAppcontainerSupported()) {
-    throw new Error(`Platform not supported: ${process.platform}.`);
+    throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 
   return new Promise((resolve, reject) => {
@@ -16,7 +29,7 @@ export function addLoopbackExemptionForAppcontainer(name: string): Promise<void>
 
     childProcess.exec(command, (error, stdout) => {
       if (error) {
-        reject(error);
+        reject(stdout);
       } else {
         resolve();
       }
@@ -29,7 +42,7 @@ export function addLoopbackExemptionForAppcontainer(name: string): Promise<void>
  * @returns True if platform supports using appcontainer; false otherwise.
  */
 export function isAppcontainerSupported() {
-  return (process.platform === "win32");
+  return process.platform === "win32";
 }
 
 /**
@@ -39,7 +52,7 @@ export function isAppcontainerSupported() {
  */
 export function isLoopbackExemptionForAppcontainer(name: string): Promise<boolean> {
   if (!isAppcontainerSupported()) {
-    throw new Error(`Platform not supported: ${process.platform}.`);
+    throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 
   return new Promise((resolve, reject) => {
@@ -47,7 +60,7 @@ export function isLoopbackExemptionForAppcontainer(name: string): Promise<boolea
 
     childProcess.exec(command, (error, stdout) => {
       if (error) {
-        reject(error);
+        reject(stdout);
       } else {
         const expr = new RegExp(`Name: ${name}`, "i");
         const found: boolean = expr.test(stdout);
@@ -75,6 +88,72 @@ export function getAppcontainerName(sourceLocation: string, isFromStore = false)
   return name;
 }
 
+export async function getUserConfirmation(name: string): Promise<boolean> {
+  const question = {
+    message: `Allow localhost loopback for ${name}?`,
+    name: "didUserConfirm",
+    type: "confirm",
+  };
+  const answers = await inquirer.prompt([question]);
+  return (answers as any).didUserConfirm;
+}
+
+export async function getAppcontainerNameFromManifestPath(manifestPath: string): Promise<string> {
+  switch (manifestPath.toLowerCase()) {
+    case "edgewebview":
+      return EdgeWebViewAppcontainerName;
+    case "edgewebbrowser":
+    case "edge":
+      return EdgeBrowserAppcontainerName;
+    default:
+      return await getAppcontainerNameFromManifest(manifestPath);
+  }
+}
+
+export function getDisplayNameFromManifestPath(manifestPath: string): string {
+  switch (manifestPath.toLowerCase()) {
+    case "edgewebview":
+      return EdgeWebViewName;
+    case "edgewebbrowser":
+    case "edge":
+      return EdgeBrowserName;
+    default:
+      return manifestPath;
+  }
+}
+
+export async function ensureLoopbackIsEnabled(
+  manifestPath: string,
+  askForConfirmation: boolean = true
+): Promise<boolean> {
+  const name = await getAppcontainerNameFromManifestPath(manifestPath);
+  let isEnabled = await isLoopbackExemptionForAppcontainer(name);
+
+  if (!isEnabled) {
+    if (!askForConfirmation || (await getUserConfirmation(getDisplayNameFromManifestPath(manifestPath)))) {
+      await addLoopbackExemptionForAppcontainer(name);
+      isEnabled = true;
+    }
+  }
+
+  return isEnabled;
+}
+
+/**
+ * Returns the name of the appcontainer used to run an Office Add-in.
+ * @param manifestPath Path of the manifest file.
+ */
+export async function getAppcontainerNameFromManifest(manifestPath: string): Promise<string> {
+  const manifest = await OfficeAddinManifest.readManifestFile(manifestPath);
+  const sourceLocation = manifest.defaultSettings ? manifest.defaultSettings.sourceLocation : undefined;
+
+  if (sourceLocation === undefined) {
+    throw new ExpectedError(`The source location could not be retrieved from the manifest.`);
+  }
+
+  return getAppcontainerName(sourceLocation, false);
+}
+
 /**
  * Removes a loopback exemption for the appcontainer.
  * @param name Appcontainer name
@@ -82,7 +161,7 @@ export function getAppcontainerName(sourceLocation: string, isFromStore = false)
  */
 export function removeLoopbackExemptionForAppcontainer(name: string): Promise<void> {
   if (!isAppcontainerSupported()) {
-    throw new Error(`Platform not supported: ${process.platform}.`);
+    throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 
   return new Promise((resolve, reject) => {
@@ -90,7 +169,7 @@ export function removeLoopbackExemptionForAppcontainer(name: string): Promise<vo
 
     childProcess.exec(command, (error, stdout) => {
       if (error) {
-        reject(error);
+        reject(stdout);
       } else {
         resolve();
       }
