@@ -1,21 +1,43 @@
-#!/usr/bin/env node
-
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-import * as commander from "commander";
 import * as fs from "fs";
+import { OfficeAddinManifest } from "office-addin-manifest";
 import * as fspath from "path";
-import * as commands from "./commands";
-import * as registry from "./dev-settings-registry";
+import * as devSettingsMac from "./dev-settings-mac";
+import * as devSettingsWindows from "./dev-settings-windows";
+import { ExpectedError } from "office-addin-usage-data";
+
+/* global process */
 
 const defaultRuntimeLogFileName = "OfficeAddins.log.txt";
 
+export { toWebViewTypeName } from "./dev-settings-windows";
+
 export enum DebuggingMethod {
+  /* eslint-disable no-unused-vars */
   Direct,
   Proxy,
   /** @deprecated use Proxy */
   Web = Proxy,
+}
+
+export enum WebViewType {
+  /* eslint-disable no-unused-vars */
+  Default = "Default",
+  IE = "IE",
+  Edge = "Edge",
+  EdgeChromium = "Edge Chromium",
+}
+
+export class RegisteredAddin {
+  public id: string;
+  public manifestPath: string;
+
+  constructor(id: string, manifestPath: string) {
+    this.id = id;
+    this.manifestPath = manifestPath;
+  }
 }
 
 export class SourceBundleUrlComponents {
@@ -25,10 +47,10 @@ export class SourceBundleUrlComponents {
   public extension?: string;
 
   public get url(): string {
-    const host = (this.host !== undefined) ? this.host : "localhost";
-    const port = (this.port !== undefined) ? this.port : "8081";
-    const path = (this.path !== undefined) ? this.path : "{path}";
-    const extension = (this.extension !== undefined) ? this.extension : ".bundle";
+    const host = this.host !== undefined ? this.host : "localhost";
+    const port = this.port !== undefined ? this.port : "8081";
+    const path = this.path !== undefined ? this.path : "{path}";
+    const extension = this.extension !== undefined ? this.extension : ".bundle";
 
     return `http://${host}${host && port ? ":" : ""}${port}/${path}${extension}`;
   }
@@ -44,9 +66,9 @@ export class SourceBundleUrlComponents {
 export async function clearDevSettings(addinId: string): Promise<void> {
   switch (process.platform) {
     case "win32":
-      return registry.clearDevSettings(addinId);
-  default:
-    throw new Error(`Platform not supported: ${process.platform}.`);
+      return devSettingsWindows.clearDevSettings(addinId);
+    default:
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
@@ -61,37 +83,42 @@ export async function disableLiveReload(addinId: string): Promise<void> {
 export async function disableRuntimeLogging(): Promise<void> {
   switch (process.platform) {
     case "win32":
-      return registry.disableRuntimeLogging();
-  default:
-    throw new Error(`Platform not supported: ${process.platform}.`);
+      return devSettingsWindows.disableRuntimeLogging();
+    default:
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
-export async function enableDebugging(addinId: string, enable: boolean = true, method: DebuggingMethod = DebuggingMethod.Direct): Promise<void> {
+export async function enableDebugging(
+  addinId: string,
+  enable: boolean = true,
+  method: DebuggingMethod = DebuggingMethod.Direct,
+  openDevTools = false
+): Promise<void> {
   switch (process.platform) {
     case "win32":
-      return registry.enableDebugging(addinId, enable, method);
+      return devSettingsWindows.enableDebugging(addinId, enable, method, openDevTools);
     default:
-      throw new Error(`Platform not supported: ${process.platform}.`);
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
 export async function enableLiveReload(addinId: string, enable: boolean = true): Promise<void> {
   switch (process.platform) {
     case "win32":
-      return registry.enableLiveReload(addinId, enable);
-  default:
-    throw new Error(`Platform not supported: ${process.platform}.`);
+      return devSettingsWindows.enableLiveReload(addinId, enable);
+    default:
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
 export async function enableRuntimeLogging(path?: string): Promise<string> {
   switch (process.platform) {
-    case "win32":
+    case "win32": {
       if (!path) {
         const tempDir = process.env.TEMP;
         if (!tempDir) {
-          throw new Error("The TEMP environment variable is not defined.");
+          throw new ExpectedError("The TEMP environment variable is not defined.");
         }
         path = fspath.normalize(`${tempDir}/${defaultRuntimeLogFileName}`);
       }
@@ -100,125 +127,158 @@ export async function enableRuntimeLogging(path?: string): Promise<string> {
       if (pathExists) {
         const stat = fs.statSync(path);
         if (stat.isDirectory()) {
-          throw new Error(`You need to specify the path to a file. This is a directory: "${path}".`);
+          throw new ExpectedError(`You need to specify the path to a file. This is a directory: "${path}".`);
         }
       }
       try {
         const file = fs.openSync(path, "a+");
         fs.closeSync(file);
       } catch (err) {
-        throw new Error(pathExists
-          ? `You need to specify the path to a writable file. Unable to write to: "${path}".`
-          : `You need to specify the path where the file can be written. Unable to write to: "${path}".`);
+        throw new ExpectedError(
+          pathExists
+            ? `You need to specify the path to a writable file. Unable to write to: "${path}".`
+            : `You need to specify the path where the file can be written. Unable to write to: "${path}".`
+        );
       }
 
-      await registry.enableRuntimeLogging(path);
+      await devSettingsWindows.enableRuntimeLogging(path);
       return path;
-  default:
-    throw new Error(`Platform not supported: ${process.platform}.`);
+    }
+    default:
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
-export async function getSourceBundleUrl(addinId: string): Promise<SourceBundleUrlComponents> {
+/**
+ * Returns the manifest paths for the add-ins that are registered
+ */
+export async function getRegisterAddIns(): Promise<RegisteredAddin[]> {
   switch (process.platform) {
+    case "darwin":
+      return devSettingsMac.getRegisteredAddIns();
     case "win32":
-      return registry.getSourceBundleUrl(addinId);
+      return devSettingsWindows.getRegisteredAddIns();
     default:
-      throw new Error(`Platform not supported: ${process.platform}.`);
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
 export async function getEnabledDebuggingMethods(addinId: string): Promise<DebuggingMethod[]> {
   switch (process.platform) {
     case "win32":
-      return registry.getEnabledDebuggingMethods(addinId);
+      return devSettingsWindows.getEnabledDebuggingMethods(addinId);
     default:
-      throw new Error(`Platform not supported: ${process.platform}.`);
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
+  }
+}
+
+export async function getOpenDevTools(addinId: string): Promise<boolean> {
+  switch (process.platform) {
+    case "win32":
+      return devSettingsWindows.getOpenDevTools(addinId);
+    default:
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
 export async function getRuntimeLoggingPath(): Promise<string | undefined> {
   switch (process.platform) {
     case "win32":
-      return registry.getRuntimeLoggingPath();
+      return devSettingsWindows.getRuntimeLoggingPath();
     default:
-      throw new Error(`Platform not supported: ${process.platform}.`);
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
+  }
+}
+
+export async function getSourceBundleUrl(addinId: string): Promise<SourceBundleUrlComponents> {
+  switch (process.platform) {
+    case "win32":
+      return devSettingsWindows.getSourceBundleUrl(addinId);
+    default:
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
+  }
+}
+
+export async function getWebView(addinId: string): Promise<WebViewType | undefined> {
+  switch (process.platform) {
+    case "win32":
+      return devSettingsWindows.getWebView(addinId);
+    default:
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
 export async function isDebuggingEnabled(addinId: string): Promise<boolean> {
   switch (process.platform) {
     case "win32":
-      return registry.isDebuggingEnabled(addinId);
+      return devSettingsWindows.isDebuggingEnabled(addinId);
     default:
-      throw new Error(`Platform not supported: ${process.platform}.`);
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
 export async function isLiveReloadEnabled(addinId: string): Promise<boolean> {
   switch (process.platform) {
     case "win32":
-      return registry.isLiveReloadEnabled(addinId);
+      return devSettingsWindows.isLiveReloadEnabled(addinId);
     default:
-      throw new Error(`Platform not supported: ${process.platform}.`);
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
+  }
+}
+
+export async function registerAddIn(manifestPath: string): Promise<void> {
+  switch (process.platform) {
+    case "win32": {
+      const manifest = await OfficeAddinManifest.readManifestFile(manifestPath);
+      const realManifestPath = fs.realpathSync(manifestPath);
+      return devSettingsWindows.registerAddIn(manifest.id || "", realManifestPath);
+    }
+    case "darwin":
+      return devSettingsMac.registerAddIn(manifestPath);
+    default:
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
 export async function setSourceBundleUrl(addinId: string, components: SourceBundleUrlComponents): Promise<void> {
   switch (process.platform) {
     case "win32":
-      return registry.setSourceBundleUrl(addinId, components);
+      return devSettingsWindows.setSourceBundleUrl(addinId, components);
     default:
-      throw new Error(`Platform not supported: ${process.platform}.`);
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
   }
 }
 
-// if this package is being run from the command line
-// (related to "main" and "bin" in package.json)
-if (process.argv[1].endsWith(fspath.join("lib", "dev-settings.js"))
-    || process.argv[1].endsWith(fspath.join(".bin", "office-addin-dev-settings"))) {
-  commander
-    .command("appcontainer <manifestPath>")
-    .description("Display or configure the appcontainer used to run the Office Add-in.")
-    .option("--loopback", `Allow access to loopback addresses such as "localhost".`)
-    .option("--prevent-loopback", `Prevent access to loopback addresses such as "localhost".`)
-    .action(commands.appcontainer);
+export async function setWebView(addinId: string, webViewType: WebViewType | undefined) {
+  switch (process.platform) {
+    case "win32":
+      return devSettingsWindows.setWebView(addinId, webViewType);
+    default:
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
+  }
+}
 
-  commander
-    .command("clear [manifestPath]")
-    .description("Clear all dev settings for the Office Add-in.")
-    .action(commands.clear);
+export async function unregisterAddIn(manifestPath: string): Promise<void> {
+  switch (process.platform) {
+    case "darwin":
+      return devSettingsMac.unregisterAddIn(manifestPath);
+    case "win32": {
+      const manifest = await OfficeAddinManifest.readManifestFile(manifestPath);
+      const realManifestPath = fs.realpathSync(manifestPath);
+      return devSettingsWindows.unregisterAddIn(manifest.id || "", realManifestPath);
+    }
+    default:
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
+  }
+}
 
-  commander
-    .command("debugging <manifest-path>")
-    .option("--enable", `Enable debugging for the add-in.`)
-    .option("--disable", "Disable debugging for the add-in")
-    .option("--debug-method <method>", "Specify the debug method: 'direct' or 'proxy'.")
-    .description("Configure debugging for the Office Add-in.")
-    .action(commands.debugging);
-
-  commander
-    .command("live-reload <manifest-path>")
-    .option("--enable", `Enable live-reload for the add-in.`)
-    .option("--disable", "Disable live-reload for the add-in")
-    .description("Configure live-reload for the Office Add-in.")
-    .action(commands.liveReload);
-
-  commander
-    .command("runtime-log")
-    .option("--enable [path]", `Enable the runtime log.`)
-    .option("--disable", "Disable the runtime log.")
-    .description("Configure the runtime log for all Office Add-ins.")
-    .action(commands.runtimeLogging);
-
-  commander
-    .command("source-bundle-url <manifestPath>")
-    .description("Specify values for components of the source bundle url.")
-    .option("-h,--host <host>", `The host name to use, or "" to use the default ('localhost').`)
-    .option("-p,--port <port>", `The port number to use, or "" to use the default (8081).`)
-    .option("--path <path>", `The path to use, or "" to use the default.`)
-    .option("-e,--extension <extension>", `The extension to use, or "" to use the default (".bundle").`)
-    .action(commands.sourceBundleUrl);
-
-  commander.parse(process.argv);
+export async function unregisterAllAddIns(): Promise<void> {
+  switch (process.platform) {
+    case "darwin":
+      return devSettingsMac.unregisterAllAddIns();
+    case "win32":
+      return devSettingsWindows.unregisterAllAddIns();
+    default:
+      throw new ExpectedError(`Platform not supported: ${process.platform}.`);
+  }
 }
