@@ -1,4 +1,8 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
 import * as winreg from "winreg";
+import { ExpectedError } from "office-addin-usage-data";
 
 export class RegistryKey {
   public winreg: winreg.Registry;
@@ -8,16 +12,23 @@ export class RegistryKey {
   }
 
   constructor(path: string) {
-    if (!path) { throw new Error("Please provide a registry key path."); }
+    if (!path) {
+      throw new ExpectedError("Please provide a registry key path.");
+    }
 
     const index = path.indexOf("\\");
 
-    if (index <= 0) { throw new Error(`The registry key path is not valid: "${path}".`); }
+    if (index <= 0) {
+      throw new ExpectedError(`The registry key path is not valid: "${path}".`);
+    }
 
     const hive = path.substring(0, index);
     const subpath = path.substring(index);
 
-    this.winreg = new winreg({ hive: normalizeRegistryHive(hive), key: subpath });
+    this.winreg = new winreg({
+      hive: normalizeRegistryHive(hive),
+      key: subpath,
+    });
   }
 }
 
@@ -57,7 +68,9 @@ async function addValue(key: RegistryKey, value: string, type: string, data: str
   return new Promise<void>((resolve, reject) => {
     const onError = (err: any) => {
       if (err) {
-        reject(new Error(`Unable to set registry value "${value}" to "${data}" (${type}) for key "${key.path}".\n${err}`));
+        reject(
+          new Error(`Unable to set registry value "${value}" to "${data}" (${type}) for key "${key.path}".\n${err}`)
+        );
       } else {
         resolve();
       }
@@ -87,19 +100,20 @@ export async function deleteKey(key: RegistryKey): Promise<void> {
   return new Promise<void>((resolve, reject) => {
     const onError = (err: any) => {
       if (err) {
-        // it's not an error if the key does not exist
-        if (err instanceof Error && err.message.match("unable to find the specified registry key")) {
-          resolve();
-        } else {
-          reject(new Error(`Unable to delete registry key "${key.path}".\n${err}`));
-        }
+        reject(new Error(`Unable to delete registry key "${key.path}".\n${err}`));
       } else {
         resolve();
       }
     };
 
     try {
-      key.winreg.destroy(onError);
+      key.winreg.keyExists((keyExistsError, exists) => {
+        if (exists) {
+          key.winreg.destroy(onError);
+        } else {
+          onError(keyExistsError);
+        }
+      });
     } catch (err) {
       onError(err);
     }
@@ -110,19 +124,20 @@ export async function deleteValue(key: RegistryKey, value: string): Promise<void
   return new Promise<void>((resolve, reject) => {
     const onError = (err: any) => {
       if (err) {
-        // it's not an error if the key or value does not exist
-        if (err instanceof Error && err.message.match("unable to find the specified registry key or value")) {
-          resolve();
-        } else {
-          reject(new Error(`Unable to delete registry value "${value}" in key "${key.path}".\n${err}`));
-        }
+        reject(new Error(`Unable to delete registry value "${value}" in key "${key.path}".\n${err}`));
       } else {
         resolve();
       }
     };
 
     try {
-      key.winreg.remove(value, onError);
+      key.winreg.valueExists(value, (_, exists) => {
+        if (exists) {
+          key.winreg.remove(value, onError);
+        } else {
+          resolve();
+        }
+      });
     } catch (err) {
       onError(err);
     }
@@ -168,17 +183,17 @@ export async function doesValueExist(key: RegistryKey, value: string): Promise<b
 export async function getNumberValue(key: RegistryKey, value: string): Promise<number | undefined> {
   const registryValue: RegistryValue | undefined = await getValue(key, value);
 
-  return (registryValue && registryValue.isNumberType) ? parseInt(registryValue.data, undefined) : undefined;
+  return registryValue && registryValue.isNumberType ? parseInt(registryValue.data, undefined) : undefined;
 }
 
 export async function getStringValue(key: RegistryKey, value: string): Promise<string | undefined> {
   const registryValue: RegistryValue | undefined = await getValue(key, value);
 
-  return (registryValue && registryValue.isStringType) ? registryValue.data : undefined;
+  return registryValue && registryValue.isStringType ? registryValue.data : undefined;
 }
 
 export async function getValue(key: RegistryKey, value: string): Promise<RegistryValue | undefined> {
-  return new Promise<RegistryValue>((resolve, reject) => {
+  return new Promise<RegistryValue | undefined>((resolve) => {
     const onError = (err: any, item?: winreg.RegistryItem) => {
       if (err) {
         resolve(undefined);
@@ -195,9 +210,27 @@ export async function getValue(key: RegistryKey, value: string): Promise<Registr
   });
 }
 
+export async function getValues(key: RegistryKey): Promise<RegistryValue[]> {
+  return new Promise<RegistryValue[]>((resolve, reject) => {
+    const callback = (err: Error, items: winreg.RegistryItem[]) => {
+      if (err) {
+        reject(err);
+      } else {
+        resolve(items.map((item) => new RegistryValue(key.path, item.name, item.type, item.value)));
+      }
+    };
+
+    try {
+      key.winreg.values(callback);
+    } catch (err) {
+      reject(err);
+    }
+  });
+}
+
 export function isNumberType(registryType: string) {
   // NOTE: REG_QWORD is not included as a number type since it cannot be returned as a "number".
-  return (registryType === RegistryTypes.REG_DWORD);
+  return registryType === RegistryTypes.REG_DWORD;
 }
 
 export function isStringType(registryType: string) {
